+++ CHANGELOG.md
@@ -12,6 +12,31 @@
 
 删除（ Removed ）：已删除的功能。
 
+
+## 1.0.0 - 2023-05-22
+
+> 简介：设计了一个简单的命令行界面，通过步骤引导完成相同格式和大小的图片集合成长图。
+
+### Features
+
+- 设计了 catalog.txt 文件读取图片名并手动处理图片顺序。
+- 修改 main() 函数为界面。
+- 添加了 argparse 模块及其部分参数的输入功能。
+- 添加了 load_catalog() 函数生成 catalog 文件。
+- 添加了 v_splice_images() 函数纵向合成图片。
+- 添加了 h_splice_images() 函数横向合成图片。
+- 添加了 load_images_path() 函数读取图片完整路径。
+- 添加了三个 get 函数作为 images_path、catalog_path 和 get_images_complete_path 的查看接口。
+
+### Refactored
+
+- 优化了文件后缀识别方法。
+- 优化了 README 内容。
+
+### Changed
+
+- load_img()函数变更为load_images_path()。
+
 ## 0.2.0 - 2023-05-04
 
 > 简介：支持部分非英文路径图片合成长图，对 PDF 导出的图片可以直接合成长图。